refactor: split watches.py into watch_fetch.py and watch_format.py
watches.py shrinks from 1060 to 584 lines by extracting:
- cogs/watch_fetch.py: all four async fetch functions (fetch_active_watches_nws,
  fetch_latest_watch_numbers, fetch_watch_details_iem, fetch_watch_details)
  plus the regex patterns and _nws_last_parsed state they share
- cogs/watch_format.py: _build_watch_embed() and _watch_files() — pure embed
  building with no I/O, independently testable

watches.py retains WatchPaginatorView, _execute_watches, WatchesCog, and
setup() — the UI/orchestration layer that depends on discord.py.

Update test_watches.py patch paths to the new module locations.
